评论统计，优化导入

+++ app/web/model/Service.py
@@ -17,7 +17,7 @@ class Service(db.Model):
     p_uid = db.Column(db.Integer, nullable=False, info='发布人id')
     category = db.Column(db.Integer, nullable=False, info='分类')
     salesVolume = db.Column(db.Integer, info='使用量')
-    score = db.Column(db.Integer, info='评分')
+    score = db.Column(db.Numeric(10, 2), info='评分')
     created = db.Column(db.DateTime, server_default=db.FetchedValue(), info='创建时间')
     updated = db.Column(db.DateTime, server_default=db.FetchedValue(), info='更新时间')
     beginDate = db.Column(db.DateTime, nullable=False, server_default=db.FetchedValue(), info='服务开始时间')
